third_party/blink/web_tests/platform/generic/fast/dom/NodeList/nodelist-item-assignment-expected.txt

This tests that indexed assignments to items of a NodeList do not override original values. See http://code.google.com/p/chromium/issues/detail?id=27967

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS first == null is false
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS nodeList[0] is first
PASS successfullyParsed is true

TEST COMPLETE
